Breakout `EmailPrompt` into separate file
